Combine voice manager directly to player
