+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "rainlink",
-  "version": "0.7.2",
+  "version": "0.7.3",
   "description": "Another lavalink wrapper but focus on stability and rich features",
   "repository": {
     "type": "git",
