+++ src/Drivers/Nodelink2.ts
@@ -112,8 +112,6 @@ export class Nodelink2 extends AbstractDriver {
 
 		const res = await fetch(url.origin + options.path, options);
 
-		// this.debug(`Request URL: ${url.origin}${options.path}`);
-
 		if (res.status == 204) {
 			this.debug('Player now destroyed');
 			return undefined;
@@ -123,6 +121,7 @@ export class Nodelink2 extends AbstractDriver {
 				'Something went wrong with lavalink server. ' +
           `Status code: ${res.status}\n Headers: ${util.inspect(options.headers)}`,
 			);
+			this.debug(`${options.method} ${options.path} ${options.body ? String(options.body) : ''}`);
 			return undefined;
 		}
 
@@ -133,7 +132,7 @@ export class Nodelink2 extends AbstractDriver {
 			finalData = this.convertV4trackResponse(finalData) as D;
 		}
 
-		this.debug(`${options.method} ${options.path}`);
+		this.debug(`${options.method} ${options.path} ${options.body ? String(options.body) : ''}`);
 
 		return finalData;
 	}
