+++ src/Manager/RainlinkPlayerManager.ts
@@ -1,13 +1,10 @@
 import { RainlinkEvents, RainlinkPlayerState, VoiceState } from '../Interface/Constants';
 import { VoiceChannelOptions } from '../Interface/Player';
-import { RainlinkVoiceManager } from './RainlinkVoiceManager';
 import { RainlinkPlayer } from '../Player/RainlinkPlayer';
 import { Rainlink } from '../Rainlink';
 import { RainlinkPlugin } from '../Plugin/VoiceReceiver/Plugin';
 
 export class RainlinkPlayerManager extends Map<string, RainlinkPlayer> {
-	/** The map of all voice manager that being handled */
-	public voiceManagers: Map<string, RainlinkVoiceManager>;
 	/** The rainlink manager */
 	public manager: Rainlink;
 
@@ -16,9 +13,8 @@ export class RainlinkPlayerManager extends Map<string, RainlinkPlayer> {
    * @param manager The rainlink manager
    * @param voiceManagers The rainlink discord voice manager class
    */
-	constructor(manager: Rainlink, voiceManagers: Map<string, RainlinkVoiceManager>) {
+	constructor(manager: Rainlink) {
 		super();
-		this.voiceManagers = voiceManagers;
 		this.manager = manager;
 	}
 
@@ -28,44 +24,35 @@ export class RainlinkPlayerManager extends Map<string, RainlinkPlayer> {
    * @internal
    */
 	async create(options: VoiceChannelOptions): Promise<RainlinkPlayer> {
-		if (this.voiceManagers.has(options.guildId))
-			throw new Error('This guild already have an existing connection');
-		const voiceManager = new RainlinkVoiceManager(this.manager, options);
-		this.voiceManagers.set(voiceManager.guildId, voiceManager);
+		if (this.has(options.guildId)) throw new Error('This guild already have an existing player');
+		const getCustomNode = this.manager.nodes.get(String(options.nodeName ? options.nodeName : ''));
+		const node = getCustomNode ? getCustomNode : await this.manager.nodes.getLeastUsed();
+		if (!node) throw new Error('Can\'t find any nodes to connect on');
+		const customPlayer =
+      this.manager.rainlinkOptions.options!.structures &&
+      this.manager.rainlinkOptions.options!.structures.player;
+		let player = customPlayer
+			? new customPlayer(this.manager, options, node)
+			: new RainlinkPlayer(this.manager, options, node);
+		this.set(player.guildId, player);
 		try {
-			await voiceManager.connect();
-		} catch (error) {
-			this.voiceManagers.delete(options.guildId);
-			throw error;
+			player = await player.connect();
+		} catch (err) {
+			this.delete(player.guildId);
+			throw err;
 		}
-		try {
-			const getCustomNode = this.manager.nodes.get(String(options.nodeName ? options.nodeName : ''));
-			const node = getCustomNode ? getCustomNode : await this.manager.nodes.getLeastUsed();
-			if (!node) throw new Error('Can\'t find any nodes to connect on');
-			const customPlayer =
-        this.manager.rainlinkOptions.options!.structures &&
-        this.manager.rainlinkOptions.options!.structures.player;
-			const player = customPlayer
-				? new customPlayer(this.manager, options, node, voiceManager)
-				: new RainlinkPlayer(this.manager, options, node, voiceManager);
-			const onUpdate = (state: VoiceState) => {
-				if (state !== VoiceState.SESSION_READY) return;
-				player.sendServerUpdate(voiceManager);
-			};
-			await player.sendServerUpdate(voiceManager);
-			voiceManager.on('connectionUpdate', onUpdate);
-			this.set(player.guildId, player);
-			player.state = RainlinkPlayerState.CONNECTED;
-			this.debug('Player created at ' + options.guildId);
-			this.manager.emit(RainlinkEvents.PlayerCreate, player);
-			const voiceReceiver = this.manager.plugins.get('rainlink-voiceReceiver') as RainlinkPlugin;
-			if (voiceManager && node.driver.id.includes('nodelink')) voiceReceiver.open(node, options);
-			return player;
-		} catch (error) {
-			voiceManager.disconnect();
-			this.voiceManagers.delete(options.guildId);
-			throw error;
-		}
+		const onUpdate = (state: VoiceState) => {
+			if (state !== VoiceState.SESSION_READY) return;
+			player.sendServerUpdate();
+		};
+		await player.sendServerUpdate();
+		player.on('connectionUpdate', onUpdate);
+		player.state = RainlinkPlayerState.CONNECTED;
+		this.debug('Player created at ' + options.guildId);
+		this.manager.emit(RainlinkEvents.PlayerCreate, player);
+		const voiceReceiver = this.manager.plugins.get('rainlink-voiceReceiver') as RainlinkPlugin;
+		if (node.driver.id.includes('nodelink')) voiceReceiver.open(node, options);
+		return player;
 	}
 
 	/**
