+++ src/Rainlink.ts
@@ -7,7 +7,6 @@ import {
 } from './Interface/Manager';
 import { EventEmitter } from 'events';
 import { RainlinkNode } from './Node/RainlinkNode';
-import { RainlinkVoiceManager } from './Manager/RainlinkVoiceManager';
 import { AbstractLibrary } from './Library/AbstractLibrary';
 import { VoiceChannelOptions } from './Interface/Player';
 import { RainlinkPlayerManager } from './Manager/RainlinkPlayerManager';
@@ -392,10 +391,6 @@ export class Rainlink extends EventEmitter {
    */
 	public readonly library: AbstractLibrary;
 	/**
-   * Voice voice managers being handled
-   */
-	public readonly voiceManagers: Map<string, RainlinkVoiceManager>;
-	/**
    * Lavalink server that has been configured
    */
 	public nodes: RainlinkNodeManager;
@@ -449,10 +444,9 @@ export class Rainlink extends EventEmitter {
       this.rainlinkOptions.options.additionalDriver?.length !== 0
 		)
 			this.drivers.push(...this.rainlinkOptions.options.additionalDriver);
-		this.voiceManagers = new Map();
 		this.nodes = new RainlinkNodeManager(this);
 		this.library.listen(this.rainlinkOptions.nodes);
-		this.players = new RainlinkPlayerManager(this, this.voiceManagers);
+		this.players = new RainlinkPlayerManager(this);
 		this.searchEngines = new Map<string, string>();
 		this.searchPlugins = new Map<string, SourceRainlinkPlugin>();
 		this.plugins = new Map<string, RainlinkPlugin>();
